changed the formatting and indentation, trimmed blank space. edited t… (#25)
changed the formatting and indentation, trimmed blank space. edited the about me text on both readme and tpl. vs code was saying that there was missing a `</p>` `</div>`tag on line 81 but I have enclosed it. added `</a>` tag in the first card graph because I still forgot it. added it in the 2 graph as well. forgot the `</div>` under the last streak stats graph. also9 added `</div>`to waktime graph on the correct nested level. and added backticks with txt to show in template that it should format as text code block. formatted the streak stats card better with help from the site.

Signed-off-by: Chunnyluny <102756485+Chunnyluny@users.noreply.github.com>